update: TimeStampCorderを別ファイルに切り分ける
