+++ TimeStampApp/Model/TimeStampRepository.swift
@@ -69,15 +69,3 @@ struct UserDefaultTimeStampRepository: TimeStampRepository {
         userDefault.timeStamps?.removeAll()
     }
 }
-
-struct TimeStampCorder {
-    func decode(data: Data?) -> [TimeStamp]? {
-        let jsonDecoder = JSONDecoder()
-        guard let data else { return nil }
-        return try? jsonDecoder.decode([TimeStamp].self, from: data)
-    }
-    
-    func encode(timeStamps: [TimeStamp]) -> Data? {
-        return try? JSONEncoder().encode(timeStamps)
-    }
-}
